budtender_chatbot/tests/test_mobile_interface.py
```python
import unittest
from unittest.mock import patch
from src import mobile_interface

class TestMobileInterface(unittest.TestCase):

    @patch('src.mobile_interface.MobileInterface')
    def setUp(self, MockMobileInterface):
        self.mobile_interface = MockMobileInterface()

    def test_speak_to_text(self):
        self.mobile_interface.speak_to_text.return_value = "Hello, how can I help you?"
        result = self.mobile_interface.speak_to_text()
        self.assertEqual(result, "Hello, how can I help you?")

    def test_text_to_speech(self):
        self.mobile_interface.text_to_speech.return_value = "Sure, let me find that for you."
        result = self.mobile_interface.text_to_speech()
        self.assertEqual(result, "Sure, let me find that for you.")

    def test_display_strain_info(self):
        strain_info = {
            "name": "Blue Dream",
            "type": "Hybrid",
            "effects": "Relaxed, Happy, Euphoric",
            "medical": "Stress, Depression, Pain",
            "flavors": "Blueberry, Berry, Sweet"
        }
        self.mobile_interface.display_strain_info.return_value = strain_info
        result = self.mobile_interface.display_strain_info()
        self.assertEqual(result, strain_info)

    def test_initiate_chat(self):
        self.mobile_interface.initiate_chat.return_value = "Chat initiated."
        result = self.mobile_interface.initiate_chat()
        self.assertEqual(result, "Chat initiated.")

    def test_end_chat(self):
        self.mobile_interface.end_chat.return_value = "Chat ended."
        result = self.mobile_interface.end_chat()
        self.assertEqual(result, "Chat ended.")

if __name__ == '__main__':
    unittest.main()
```